# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: Polla Mundial 2026 API
  version: 0.1.0
  description: Backend privado de pronósticos para el Mundial 2026. ~20 participantes por invitación.
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://api.polla2026.com
    description: Producción (TBD)
...
security:
  - cookieAuth: []
paths:
...
  /admin/invitations:
    post:
      tags: [admin]
      summary: Genera un código de invitación de un solo uso
      operationId: postAdminInvitations
      description: >
        Crea un único código alfanumérico aleatorio con validez de 24 horas.
        Solo accesible por admin.
      responses:
        '201':
          description: Código generado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
              example:
                id: "uuid"
                code: "XKMP1234"
                status: "AVAILABLE"
                usedAt: null
                expiresAt: "2026-06-12T15:00:00.000Z"
                createdAt: "2026-06-11T15:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [admin]
      summary: Lista todos los códigos de invitación
      operationId: getAdminInvitations
      description: Soporta filtro opcional por estado y paginación. Solo accesible por admin.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [AVAILABLE, USED, EXPIRED]
          description: "Filtra por estado. Si se omite retorna todos."
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: "Número de página (base 1)."
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
          description: "Cantidad de resultados por página."
      responses:
        '200':
          description: Lista paginada de invitaciones
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, pageSize]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invitation'
                  total:
                    type: integer
                    description: "Total de invitaciones que coinciden con el filtro."
                  page:
                    type: integer
                    description: "Página actual."
                  pageSize:
                    type: integer
                    description: "Tamaño de página usado."
              example:
                total: 2
                page: 1
                pageSize: 20
                data:
                  - { id: "uuid1", code: "XKMP1234", status: "USED",      usedAt: "2026-06-11T16:00:00Z", expiresAt: "2026-06-12T15:00:00Z", createdAt: "2026-06-11T15:00:00Z" }
                  - { id: "uuid2", code: "ABNQ5678", status: "AVAILABLE", usedAt: null,                   expiresAt: "2026-06-12T16:00:00Z", createdAt: "2026-06-11T16:00:00Z" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/groups:
    post:
      tags: [admin]
      summary: Carga los 12 grupos del torneo con sus equipos
      operationId: postAdminGroups
      description: >
        Operación de una sola vez antes del inicio del torneo.
        Acepta un array de 12 grupos, cada uno con sus 4 equipos.
        Falla si los grupos ya existen.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groups]
              properties:
                groups:
                  type: array
                  items:
                    $ref: '#/components/schemas/GroupInput'
                  minItems: 12
                  maxItems: 12
      responses:
        '201':
          description: Grupos y equipos cargados correctamente
          content:
            application/json:
              schema:
                type: object
                required: [ok, groupsCount, teamsCount]
                properties:
                  ok:          { type: boolean }
                  groupsCount: { type: integer }
                  teamsCount:  { type: integer }
              example: { ok: true, groupsCount: 12, teamsCount: 48 }
        '409':
          description: Los grupos ya fueron cargados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { code: "GROUPS_ALREADY_LOADED", message: "Groups have already been loaded" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/ko/matches:
    post:
      tags: [admin]
      summary: Carga los partidos de una ronda KO
      operationId: postAdminKoMatches
      description: >
        Se llama ronda a ronda una vez que los cruces son oficiales.
        Requiere el slug de la ronda y un array de partidos.
        Cada partido incluye externalMatchId (ID 1–104 de la API externa worldcup2026)
        para poder consultar resultados en tiempo real.
        lockedAt se calcula automáticamente como scheduledAt - 30 minutos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roundSlug, matches]
              properties:
                roundSlug:
                  type: string
                  enum: [GROUP, R32, R16, QF, SF, THIRD, FINAL]
                matches:
                  type: array
                  items:
                    $ref: '#/components/schemas/KoMatchInput'
                  minItems: 1
            example:
              roundSlug: "R32"
              matches:
                - externalMatchId: 73
                  matchNumber: 1
                  homeTeamId: null
                  awayTeamId: null
                  homeTeamLabel: "Runner-up Group A"
                  awayTeamLabel: "Runner-up Group B"
                  scheduledAt: "2026-06-28T12:00:00Z"
                - externalMatchId: 74
                  matchNumber: 2
                  homeTeamId: "uuid-col"
                  awayTeamId: "uuid-bra"
                  homeTeamLabel: null
                  awayTeamLabel: null
                  scheduledAt: "2026-06-28T16:00:00Z"
      responses:
        '201':
          description: Partidos cargados correctamente
          content:
            application/json:
              schema:
                type: object
                required: [ok, roundSlug, matchesCount]
                properties:
                  ok:           { type: boolean }
                  roundSlug:    { type: string }
                  matchesCount: { type: integer }
              example: { ok: true, roundSlug: "R32", matchesCount: 16 }
        '404':
          description: Ronda no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { code: "ROUND_NOT_FOUND", message: "Round 'QF' not found" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/ko/matches/{matchId}/result:
    put:
      tags: [admin]
      summary: Carga o corrige el resultado oficial de un partido KO
      operationId: putAdminKoMatchResult
      description: >
        Registra el marcador a 120' y el equipo clasificado.
        winnerTeamId es siempre obligatorio para evitar ambigüedad en casos
        de penales donde el marcador no determina el ganador.
        Tras guardar el resultado, el sistema actualiza status a finished.
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scoreHome, scoreAway, winnerTeamId]
              properties:
                scoreHome:    { type: integer, minimum: 0 }
                scoreAway:    { type: integer, minimum: 0 }
                winnerTeamId: { type: string, format: uuid }
            example:
              scoreHome: 1
              scoreAway: 1
              winnerTeamId: "uuid-col"
      responses:
        '200':
          description: Resultado registrado correctamente
          content:
            application/json:
              schema:
                type: object
                required: [ok, matchId, scoreHome, scoreAway, winnerTeamId]
                properties:
                  ok:           { type: boolean }
                  matchId:      { type: string, format: uuid }
                  scoreHome:    { type: integer }
                  scoreAway:    { type: integer }
                  winnerTeamId: { type: string, format: uuid }
              example: { ok: true, matchId: "uuid", scoreHome: 1, scoreAway: 1, winnerTeamId: "uuid-col" }
        '404':
          description: Partido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { error: "Partido no encontrado", code: "MATCH_NOT_FOUND" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/scoring-params/{key}:
    put:
      tags: [admin]
      summary: Actualiza el valor de un parámetro de puntuación
      operationId: putAdminScoringParams
      description: >
        Edita un parámetro por su key. El recálculo de puntajes
        se aplica en la siguiente corrida del cron o al consultar el scoreboard.
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            enum:
              - pts_group_position_exact
              - pts_group_position_partial
              - bonus_group_complete
              - pts_third_correct
              - pts_ko_advances
              - pts_ko_exact_score
              - pts_dark_horse_per_round
              - pts_disappointment_per_round
              - mult_triple
              - scale_r32
              - scale_r16
              - scale_qf
              - scale_sf
              - scale_final
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: number }
            example: { value: 5 }
      responses:
        '200':
          description: Parámetro actualizado
          content:
            application/json:
              schema:
                type: object
                required: [ok, key, value, updatedAt]
                properties:
                  ok:        { type: boolean }
                  key:       { type: string }
                  value:     { type: number }
                  updatedAt: { type: string, format: date-time }
              example: { ok: true, key: "pts_ko_exact_score", value: 5, updatedAt: "2026-06-11T15:00:00Z" }
        '404':
          description: Key no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { error: "Parámetro no encontrado", code: "PARAM_NOT_FOUND" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/top8:
    put:
      tags: [admin]
      summary: Define los 8 equipos top FIFA
      operationId: putAdminTop8
      description: >
        Foto fija para todo el torneo. Marca como isTop8 true los equipos
        del array y false al resto. Se puede llamar varias veces para corregir.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamIds]
              properties:
                teamIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 8
                  maxItems: 8
            example:
              teamIds:
                - "uuid-bra"
                - "uuid-fra"
                - "uuid-eng"
                - "uuid-arg"
                - "uuid-esp"
                - "uuid-por"
                - "uuid-bel"
                - "uuid-ned"
      responses:
        '200':
          description: Top 8 actualizado correctamente
          content:
            application/json:
              schema:
                type: object
                required: [ok, teams]
                properties:
                  ok: { type: boolean }
                  teams:
                    type: array
                    items:
                      type: object
                      required: [id, name, code]
                      properties:
                        id:   { type: string, format: uuid }
                        name: { type: string }
                        code: { type: string }
                        flag: { type: [string, 'null'], description: "URL de la bandera del equipo" }
        '400':
          description: No se enviaron exactamente 8 IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { code: "INVALID_TOP8_COUNT", message: "Exactly 8 team IDs required" }
        '404':
          description: Uno o más IDs de equipo no existen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { code: "TEAM_NOT_FOUND", message: "One or more team IDs not found" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/participants:
    get:
      tags: [admin]
      summary: Lista participantes con puntaje actual y datos de contacto
      operationId: getAdminParticipants
      responses:
        '200':
          description: Lista de participantes
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminParticipant'
              example:
                data:
                  - id: "uuid-juan"
                    name: "Juan"
                    email: "juan@gmail.com"
                    phone: "+573001234567"
                    role: "PARTICIPANT"
                    totalScore: 145
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/groups/thirds:
    put:
      tags: [admin]
      summary: Registra los 8 mejores terceros oficiales del torneo
      operationId: putAdminGroupsThirds
      description: >
        El admin marca cuáles 8 equipos clasificaron como mejores terceros oficialmente.
        Todos los teamIds deben ser equipos que hayan quedado en posición 3 real según
        los standings del torneo. Dispara el recálculo de puntos por terceros.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamIds]
              properties:
                teamIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 8
                  maxItems: 8
            example:
              teamIds:
                - "uuid-col"
                - "uuid-uru"
                - "uuid-sen"
                - "uuid-pol"
                - "uuid-cro"
                - "uuid-aus"
                - "uuid-mar"
                - "uuid-irn"
      responses:
        '200':
          description: Terceros registrados correctamente
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
              example: { ok: true }
        '400':
          description: teamId no válido o cantidad incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                cantidad_incorrecta:
                  value: { code: "INVALID_THIRDS_COUNT", message: "Exactly 8 team IDs required" }
                equipo_invalido:
                  value: { code: "INVALID_THIRD_TEAM", message: "Team uuid-xxx is not a third-place team" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/groups/{groupId}/locked:
    put:
      tags: [admin]
      summary: Abre o cierra un grupo para predicciones
      operationId: putAdminGroupLocked
      description: >
        Marca un grupo específico como cerrado (`locked: true`) o abierto (`locked: false`).
        Cuando está cerrado, cualquier intento de guardar predicciones para ese grupo
        retorna 423 GROUP_LOCKED. Independiente del candado global de la fase de grupos.
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locked]
              properties:
                locked: { type: boolean }
            examples:
              cerrar:
                summary: Cerrar el grupo
                value: { locked: true }
              abrir:
                summary: Volver a abrir el grupo
                value: { locked: false }
      responses:
        '200':
          description: Estado actualizado correctamente
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
              example: { ok: true }
        '404':
          description: Grupo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { code: "GROUP_NOT_FOUND", message: "Group not found" }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /groups:
    get:
      tags: [groups]
      summary: Lista los 12 grupos con sus equipos
      operationId: getGroups
      responses:
        '200':
          description: Lista de grupos
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
              example:
                data:
                  - id: "uuid-A"
                    label: "A"
                    name: "Grupo A"
                    locked: false
                    teams:
                      - id: "uuid-mex"
                        name: "México"
                        code: "MEX"
                        isTop8: false
                        flag: "https://flagcdn.com/w80/mx.png"
                        positionStats:
                          - { position: 1, pct: 60.0 }
                          - { position: 2, pct: 25.0 }
                          - { position: 3, pct: 15.0 }
                      - id: "uuid-usa"
                        name: "Estados Unidos"
                        code: "USA"
                        isTop8: true
                        flag: "https://flagcdn.com/w80/us.png"
                        positionStats: null
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /groups/thirds:
    get:
      tags: [groups]
      summary: Candidatos a mejores terceros del participante autenticado
      operationId: getGroupsThirds
      description: >
        Retorna los equipos que el participante predijo en posición 3,
        uno por grupo. Solo aparecen grupos con predicción completa.
        El flag selected indica si ya fue marcado como mejor tercero
        en POST /groups/thirds.
      responses:
        '200':
          description: Lista de candidatos a tercero
          content:
            application/json:
              schema:
                type: object
                required: [data, selectedCount]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ThirdCandidate'
                  selectedCount:
                    type: integer
                    description: "Cuántos de los candidatos están marcados como mejor tercero (máx 8)"
              example:
                selectedCount: 2
                data:
                  - { teamId: "uuid-col", name: "Colombia", code: "COL", flag: "https://flagcdn.com/w80/co.png", groupId: "uuid-A", label: "A", selected: true  }
                  - { teamId: "uuid-uru", name: "Uruguay",  code: "URU", flag: "https://flagcdn.com/w80/uy.png", groupId: "uuid-B", label: "B", selected: false }
                  - { teamId: "uuid-sen", name: "Senegal",  code: "SEN", flag: "https://flagcdn.com/w80/sn.png", groupId: "uuid-C", label: "C", selected: true  }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [groups]
      summary: Guarda o reemplaza la selección de mejores terceros
      operationId: postGroupsThirds
      description: >
        Recibe exactamente 8 teamId. Todos deben ser equipos que el participante
        predijo en posición 3 (candidatos de GET /groups/thirds).
        Reemplaza cualquier selección previa (upsert completo).
        Valida candado (antes del 11 de junio de 2026).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamIds]
              properties:
                teamIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 8
                  maxItems: 8
            example:
              teamIds:
                - "uuid-col"
                - "uuid-uru"
                - "uuid-sen"
                - "uuid-pol"
                - "uuid-cro"
                - "uuid-aus"
                - "uuid-mar"
                - "uuid-irn"
      responses:
        '200':
          description: Mejores terceros guardados correctamente
          content:
            application/json:
              schema:
                type: object
                required: [ok, selectedCount]
                properties:
                  ok:            { type: boolean }
                  selectedCount: { type: integer }
              example: { ok: true, selectedCount: 8 }
        '400':
          description: >
            Payload inválido. Posibles causas: no son exactamente 8 equipos,
            algún teamId no es candidato válido (no fue predicho en posición 3).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_es_candidato:
                  value: { error: "El equipo uuid-bra no es un tercero válido en tus predicciones", code: "INVALID_THIRD_CANDIDATE" }
                cantidad_incorrecta:
                  value: { error: "Debes seleccionar exactamente 8 equipos", code: "INVALID_THIRDS_COUNT" }
        '423':
          description: Candado activo — predicciones de grupos cerradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { error: "Las predicciones de grupos están cerradas", code: "PREDICTIONS_LOCKED" }
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── KO ──────────────────────────────────────────────────────────────────────
  /ko/matches:
    get:
      tags: [ko]
      summary: Lista los partidos de una ronda KO
      operationId: getKoMatches
      description: >
        Requiere el parámetro de query roundSlug. Retorna los partidos de esa
        ronda con resultado real (si existe) y la predicción propia del
        participante autenticado por partido.
        Si un partido tiene status finished o su scheduledAt + 90 min ya pasó
        sin resultado, el BE consulta la API externa worldcup2026, guarda el
        resultado y lo retorna.
      parameters:
        - name: roundSlug
          in: query
          required: true
          schema:
            type: string
            enum: [GROUP, R32, R16, QF, SF, THIRD, FINAL]
          description: "Ronda a consultar"
      responses:
        '200':
          description: Partidos de la ronda
          content:
            application/json:
              schema:
                type: object
                required: [round, matches]
                properties:
                  round:
                    $ref: '#/components/schemas/KoRound'
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/KoMatch'
              example:
                round: { slug: "R32", name: "Dieciseisavos", order: 1 }
                matches:
                  - id: "uuid-match-1"
                    externalMatchId: 73
                    matchNumber: 1
                    scheduledAt: "2026-06-28T12:00:00Z"
                    lockedAt: "2026-06-28T11:30:00Z"
                    status: "SCHEDULED"
                    homeTeam: null
                    awayTeam: null
                    homeTeamLabel: "Runner-up Group A"
                    awayTeamLabel: "Runner-up Group B"
                    result: null
                    myPrediction: null
                  - id: "uuid-match-2"
                    externalMatchId: 74
                    matchNumber: 2
                    scheduledAt: "2026-06-28T16:00:00Z"
                    lockedAt: "2026-06-28T15:30:00Z"
                    status: "FINISHED"
                    homeTeam: { id: "uuid-col", name: "Colombia", code: "COL", flag: "https://flagcdn.com/w80/co.png" }
                    awayTeam: { id: "uuid-bra", name: "Brasil",   code: "BRA", flag: "https://flagcdn.com/w80/br.png" }
                    homeTeamLabel: null
                    awayTeamLabel: null
                    result: { scoreHome: 1, scoreAway: 1, winnerTeamId: "uuid-col" }
                    myPrediction:
                      scoreHome: 2
                      scoreAway: 0
                      teamAdvancesId: "uuid-col"
                      tripleActive: false
                      lockedIn: true
        '400':
          description: roundSlug inválido o faltante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { error: "roundSlug es requerido", code: "VALIDATION_ERROR" }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Cookie HttpOnly `session` emitida por POST /auth/google.
        JWT firmado con JWT_SECRET, expira en 30 días.
        El navegador la envía automáticamente — no requiere header explícito.
  responses:
    Unauthorized:
      description: Cookie de sesión ausente o inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            sin_cookie:
              value: { code: "MISSING_SESSION", message: "Sesión requerida" }
            cookie_invalida:
              value: { code: "INVALID_SESSION", message: "Sesión inválida o expirada" }
    Forbidden:
      description: El usuario autenticado no tiene rol admin
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { code: "FORBIDDEN", message: "Acceso denegado" }
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Recurso no encontrado
...
    Invitation:
      type: object
      required: [id, code, status, expiresAt, createdAt]
      properties:
        id:        { type: string, format: uuid }
        code:      { type: string }
        status:    { type: string, enum: [AVAILABLE, USED, EXPIRED] }
        usedAt:    { type: [string, 'null'], format: date-time }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
    GroupInput:
      type: object
      required: [label, name, teams]
      properties:
        label: { type: string, description: "Letra del grupo, ej: A" }
        name:  { type: string, description: "Nombre completo, ej: Grupo A" }
        lastMatchAt:
          type: [string, 'null']
          format: date-time
          description: "Fecha del último partido del grupo. Opcional."
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamInput'
          minItems: 4
          maxItems: 4
    KoMatchInput:
      type: object
      required: [externalMatchId, matchNumber, scheduledAt]
      description: >
        Partido de fase KO. externalMatchId corresponde al ID numérico (1–104)
        de la API externa (worldcup2026), usado para consultar resultados en tiempo real.
        homeTeamId y awayTeamId son opcionales cuando los cruces aún no se conocen;
        en ese caso se usan homeTeamLabel y awayTeamLabel como placeholders.
      properties:
        externalMatchId:
          type: integer
          description: "ID numérico de la API externa worldcup2026 (1–104)"
        matchNumber:
          type: integer
          description: "Número de partido dentro de la ronda (1-based)"
        homeTeamId:
          type: [string, 'null']
          format: uuid
        awayTeamId:
          type: [string, 'null']
          format: uuid
        homeTeamLabel:
          type: [string, 'null']
          description: "Placeholder cuando el equipo aún no se conoce, ej: Runner-up Group A"
        awayTeamLabel:
          type: [string, 'null']
          description: "Placeholder cuando el equipo aún no se conoce, ej: Runner-up Group B"
        scheduledAt:
          type: string
          format: date-time
...
    KoRound:
      type: object
      required: [slug, name, order]
      properties:
        slug:  { type: string, enum: [GROUP, R32, R16, QF, SF, THIRD, FINAL] }
        name:  { type: string }
        order: { type: integer }
...